change jog logic to float

+++ src/emc/motion/motion.h
@@ -483,7 +483,7 @@ Suggestion: Split this in to an Error and a Status flag register..
 
 	double motor_offset;	/* diff between internal and motor pos, used
 				   to set position to zero during homing */
-	int old_jjog_counts;	/* prior value, used for deltas */
+	double old_jjog_counts;	/* prior value, used for deltas */
 	double big_vel;		/* used for "debouncing" velocity */
     } emcmot_joint_t;
